+++ ichnaea/__init__.py
@@ -3,8 +3,15 @@ import os.path
 import sys
 import warnings
 
+from shapely import speedups
+
 warnings.simplefilter('ignore', DeprecationWarning)
 warnings.simplefilter('ignore', PendingDeprecationWarning)
+
+# Enable shapely speedups
+if speedups.available:
+    speedups.enable()
+del speedups
 
 # Enable pyopenssl based SSL stack
 if sys.version_info < (3, 0):  # pragma: no cover